requirements.txt
requests
Pillow
tkinterdnd2
